+++ resources/Views/auth/home.php
@@ -78,7 +78,7 @@ if (Auth::hasRole('Maintainer')) { ?>
 
 <div class="box">
     <strong><?= __('DeliveryNote') ?></strong><br>
-    <!-- <a href="showDeliveryNote_OpenVolumeDistribution">Offene Mengenverteilungen</a><br> -->
+    <a href="openVolumeDistributions"><?= __('OpenVolumeDistributions') ?></a><br>
     <a href="showVolumeDistribution"><?= __('ShowVolumeDistribution') ?></a>
 </div>